feat: add initial FeedView component
Add the first FeedView react component.

We decided to use `mantine` as our component library.

We've got the basic UX for submitting the form in place. We now just
need to actually query the backend. We want to extract the code for
querying the backend to a separate library.

We also:

- Setup the templating to `mantine`.
- Added `prettier` for formatting.
- Removed a lot of `vite` template stuff we don't need.